commit
commit

+++ com-cg-sprint-school-administration/src/main/java/com/cg/sprint/school/administration/model/Complaint.java
@@ -54,7 +54,7 @@ public class Complaint {
 		this.complaintId = complaintId;
 	}
 
-	public long getComplaintId() {
+	public int getComplaintId() {
 		return complaintId;
 	}
